requirements.txt
detoxify 
fastapi 
torch 
transformers
pillow
uvicorn